Final Dashboard
Image of the final solution and design of the dashboard. I used our software's page builder to override existing CSS classes to make it look the way I wanted it to from the wireframes. I also provided a mobile view as well. This is all shown on my project page that I provided a link to in my description.